+++ CHANGELOG.md
@@ -49,9 +49,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - Remove `rust-version` from crate manifests ([#3000](https://github.com/coral-xyz/anchor/pull/3000)).
 - cli: Fix upgradeable program clones ([#3010](https://github.com/coral-xyz/anchor/pull/3010)).
 - ts: Fix using IDLs that have defined types as generic arguments ([#3016](https://github.com/coral-xyz/anchor/pull/3016)).
-
-### Breaking
-- lang: Make program ID const instead of static ([#3019](https://github.com/coral-xyz/anchor/pull/3019)).
+- lang: Add `const` of program ID to `declare_id!` and `declare_program!` ([#3019](https://github.com/coral-xyz/anchor/pull/3019)).
 
 ## [0.30.0] - 2024-04-15